requirements.txt
uvicorn==0.19.0
fastapi==0.87.0
python-multipart
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0

### in case you use yolov8
ultralytics==8.1.34